Kill some horribly ugly hackery :( It's just not right to pass entire IniFile objects across DLLs. Only POD (plain-old-data, like int, float, char, char[] etc) are safe to pass across DLL boundaries.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4257 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/Core.cpp
@@ -705,7 +705,7 @@ void Callback_VideoCopiedToXFB(bool video_update)
 		double wait_frametime = (1000.0 / targetfps);
 
 		if (Timer.GetTimeDifference() >= wait_frametime * frames)
-			no_framelimit=Timer.GetTimeDifference();
+			no_framelimit = (u32)Timer.GetTimeDifference();
 
 		while (Timer.GetTimeDifference() < wait_frametime * frames)
 		{
@@ -756,7 +756,7 @@ void Callback_VideoCopiedToXFB(bool video_update)
 
 		int TargetVPS = (int)(VideoInterface::TargetRefreshRate + 0.5);
 
-		float Speed = ((VPS > 0 ? VPS : VideoInterface::ActualRefreshRate) / TargetVPS) * 100.0f;
+		float Speed = ((VPS > 0.0f ? VPS : VideoInterface::ActualRefreshRate) / TargetVPS) * 100.0f;
 		
 		// Settings are shown the same for both extended and summary info
 		std::string SSettings = StringFromFormat(" | Core: %s %s",
